+++ src/gemseo_umdo/formulations/_functions/__init__.py
@@ -12,17 +12,22 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-"""Factory of U-MDO formulations."""
+"""Functions to estimate statistics from an
+[UMDOFormulation][gemseo_umdo.formulations.formulation.UMDOFormulation].
+
+The base function is a
+[StatisticFunction][gemseo_umdo.formulations._functions.statistic_function.StatisticFunction]
+and derives from an
+[MDOFunction][gemseo.core.mdofunctions.mdo_function.MDOFunction].
+Most of the other _functions derive from
+[StatisticFunction][gemseo_umdo.formulations._functions.statistic_function.StatisticFunction]
+and are associated with an
+[UMDOFormulation][gemseo_umdo.formulations.formulation.UMDOFormulation],
+e.g.
+[Sampling][gemseo_umdo.formulations.sampling.Sampling].
+and
+[TaylorPolynomial][gemseo_umdo.formulations.taylor_polynomial.TaylorPolynomial].
+The other modules are helpers.
+"""  # noqa: D205
 
 from __future__ import annotations
-
-from gemseo.formulations.factory import MDOFormulationFactory
-
-from gemseo_umdo.formulations.formulation import UMDOFormulation
-
-
-class UMDOFormulationsFactory(MDOFormulationFactory):
-    """The factory of U-MDO formulations."""
-
-    _CLASS = UMDOFormulation
-    _MODULE_NAMES = ("gemseo_umdo.formulations",)