+++ src/gemseo_umdo/formulations/_statistics/sampling/standard_deviation.py
@@ -12,17 +12,20 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-"""Factory of U-MDO formulations."""
+"""Estimator of the standard deviation for sampling-based U-MDO formulations."""
 
 from __future__ import annotations
 
-from gemseo.formulations.factory import MDOFormulationFactory
+from typing import TYPE_CHECKING
 
-from gemseo_umdo.formulations.formulation import UMDOFormulation
+from gemseo_umdo.formulations._statistics.sampling.variance import Variance
+
+if TYPE_CHECKING:
+    from gemseo.typing import RealArray
 
 
-class UMDOFormulationsFactory(MDOFormulationFactory):
-    """The factory of U-MDO formulations."""
+class StandardDeviation(Variance):
+    """Estimator of the standard deviation."""
 
-    _CLASS = UMDOFormulation
-    _MODULE_NAMES = ("gemseo_umdo.formulations",)
+    def _compute(self, samples: RealArray) -> RealArray:  # noqa: D102
+        return super()._compute(samples) ** 0.5